+++ lib/src/procreate_v1/decode.dart
@@ -1,12 +1,12 @@
 part of 'procreate.dart';
 
-List<ProcreateSwatches> decodeProcreateSwatches(List<int> bytes) {
+List<ProcreateV1Palette> decodeProcreateV1Palettes(List<int> bytes) {
   final archive = ZipDecoder().decodeBytes(bytes);
   final jsonFile = archive.first;
   final jsonContent = String.fromCharCodes(jsonFile.content as List<int>);
 
   final swatches = json.decode(jsonContent) as List<dynamic>;
   return swatches.map((swatch) {
-    return ProcreateSwatchesMapper.fromMap(swatch as Map<String, dynamic>);
+    return ProcreateV1PaletteMapper.fromMap(swatch as Map<String, dynamic>);
   }).toList();
 }