+++ lib/src/procreate_v1/encode.dart
@@ -1,6 +1,6 @@
 part of 'procreate.dart';
 
-List<int> encodeProcreateSwatches(List<ProcreateSwatches> swatches) {
+List<int> encodeProcreateV1Palette(List<ProcreateV1Palette> swatches) {
   final jsonList =
       swatches.map((swatch) {
         return swatch.toMap();