+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Breaking Changes
 - Renamed Jasc format to Paint Shop Pro.
+- Renamed Procreate Swatches format to Procreate V1 Palette.
 - Reworked GIMP palette API.
 - Now targeting Dart >= 3.7 and Flutter >= 3.29.